Создал директорию для интеграционных тестов
